Editor and tests code motion done
